Smart-Task-Analyzer/backend/requirements.txt
django==4.2
djangorestframework==3.15
django-cors-headers==4.4.0
